GIS-WA2 council fixes: 3D layer drape legibility, honesty, robustness
LLM council (visual + code review) on the just-shipped 3D layer drape. Fixes the
unanimous, clearly-correct findings (broader items -- empty-states, freshness stamps,
hazard-render unification, a workarea.png DoS cache -- banked as flagged follow-ups):

- Wireframe drowned the layer (both visual reviewers' #1): hoist the wire style to one
  WIRE const, lower base opacity 0.28->0.14, and DIM to 0.05 when a real layer texture
  is draped so the layer is the dominant, decodable read. Verified: slope now reads.
- No legend / silent exaggeration: a 3D legend (ramp -> values per kind) + an explicit
  '↕×N vert. exag' / '↕ true scale' disclosure on the toolbar (precision-honesty -- the
  relief is never silently scaled). Verified live: 'slope 0°–30°+ · ↕×4.0 vert. exag'.
- setLayer had no texture-error path: add onError -> revert to height + onLayerError
  callback so the cockpit resets the dropdown + warns (no stale/wrong drape on a 404).
- Dead .astype('uint8') in render_globe (helper already returns uint8) removed.

Council confirmed correct (no change): orientation/registration (raster flipud <-> canvas
<-> 3D uv+flipY), the _layer_rgba refactor fidelity, heightfield<->png cell agreement,
sun-azimuth convention, planExtent seed. No path traversal / XSS; kind/site/window bounded.

+++ stewie/server/gis_layers.py
@@ -314,7 +314,7 @@ def render_globe(kind: str, *, sun_el: float = 6.0, sun_az: float = 90.0, mp=Non
         rgba = _layer_rgba(dem, cm, kind, sun_az, sun_el)
         if rgba is None:
             return None
-        out = _reproject(rgba.astype("uint8"), b, fwd, out_px=1024)
+        out = _reproject(rgba, b, fwd, out_px=1024)   # _layer_rgba already returns uint8
     _GLOBE_CACHE[key] = out
     # RC-03 (audit 2026-06-11): write ATOMICALLY (.part -> os.replace) so a concurrent reader /
     # the startup warm thread never sees a torn .npy; the JSON sidecar lands LAST as the commit
